+++ pkg/streamer/super_node_streamer.go
@@ -21,7 +21,7 @@ import (
 	"github.com/ethereum/go-ethereum/rpc"
 
 	"github.com/vulcanize/ipfs-blockchain-watcher/pkg/core"
-	"github.com/vulcanize/ipfs-blockchain-watcher/pkg/watcher"
+	"github.com/vulcanize/ipfs-blockchain-watcher/pkg/watch"
 )
 
 // SuperNodeStreamer is the underlying struct for the shared.SuperNodeStreamer interface
